chore: 조회 반환 세부 변경

+++ src/main/java/com/trustping/repository/AbnormalDataRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.trustping.entity.AbnormalData;
 
 public interface AbnormalDataRepository extends JpaRepository<AbnormalData, Long>{
-	public int findSACLByCarIdAndCreatedAt(Long carId, LocalDateTime createdAt);
+	public Long findSACLByCarIdAndCreatedAt(Long carId, LocalDateTime createdAt);
 }
