+++ src/main/java/com/trustping/service/AbnormalDataService.java
@@ -3,5 +3,5 @@ package com.trustping.service;
 import java.time.LocalDateTime;
 
 public interface AbnormalDataService {
-	int findSACLByCarIdAndCreatedAt(Long carId, LocalDateTime createdAt);
+	Long findSACLByCarIdAndCreatedAt(Long carId, LocalDateTime createdAt);
 }
